cv-generator-gcp/requirements.txt
functions-framework
docxtpl
requests
google-cloud-storage
google-cloud-secret-manager
pydantic
jsonschema
python-docx
Jinja2
MarkupSafe